# Backend/docs/swagger.yaml
definitions:
...
  controller.passwordResponse:
    properties:
      message:
        default: Password updated successfully
        type: string
      success:
        default: true
        type: boolean
    type: object
  controller.profileResponse:
    properties:
      data:
        $ref: '#/definitions/domain.StaffUser'
      success:
        default: true
        type: boolean
    type: object
...
  controller.updatePasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  controller.updateProfileRequest:
    properties:
      first_name:
        type: string
      last_name:
        type: string
    required:
    - first_name
    - last_name
    type: object
...
  domain.ErrorDetails:
    properties:
      code:
        type: string
      details:
        type: string
    type: object
  domain.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/domain.ErrorDetails'
      message:
        default: Error
        type: string
      success:
        default: false
        type: boolean
    type: object
...
  domain.StaffUser:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_login_at:
        type: string
      last_name:
        type: string
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/admin/me:
    get:
      description: Get detailed profile information of the authenticated staff member
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.profileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - profile
    put:
      consumes:
      - application/json
      description: Update the first and last name of the authenticated staff member
      parameters:
      - description: Profile update payload
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/controller.updateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.profileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - profile
  /api/admin/me/password:
    put:
      consumes:
      - application/json
      description: Change the password of the authenticated staff member
      parameters:
      - description: Password update payload
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/controller.updatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.passwordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user password
      tags:
      - profile
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"